### 2021-11-21 #3938 @whedon
[PRE REVIEW]: sweater: Speedy Word Embedding Association Test and Extras Using R
**Submitting author:** @chainsawriot (<a href="http://orcid.org/0000-0002-6232-7530">Chung-hong Chan</a>)
**Repository:** <a href="https://github.com/chainsawriot/sweater" target ="_blank">https://github.com/chainsawriot/sweater</a>
**Version:** v0.1.0
