### 2021-11-21 79c1549
paper: update dois; rm trash
