+++ paper/paper.bib
@@ -3,6 +3,7 @@
   author = {Jisun An and Haewoon Kwak and Yong-Yeol Ahn},
   journal = {arXiv preprint arXiv:1806.05521},
   year = {2018},
+  doi={10.18653/v1/p18-1228},
 }
 
 @InProceedings{antoniak2021bad,
@@ -11,6 +12,7 @@
   booktitle = {{Proceedings of the 59th Annual Meeting of the Association for Computational Linguistics and the 11th International Joint Conference on Natural Language Processing (Volume 1: Long Papers)}},
   pages = {1889--1904},
   year = {2021},
+  doi={10.18653/v1/2021.acl-long.148}
 }
 
 @Article{arendt:2013:DDM,
@@ -36,6 +38,7 @@
   booktitle = {IJCAI},
   pages = {430--436},
   year = {2020},
+  doi={10.24963/ijcai.2020/60}
 }
 
 @Article{boyarskaya2020overcoming,
@@ -99,6 +102,7 @@
   booktitle = {{Proceedings of the 30th ACM conference on hypertext and social media}},
   pages = {253--257},
   year = {2019},
+  doi={10.1145/3342220.3343658}
 }
 
 @Article{kroon2020guilty,
@@ -116,6 +120,7 @@
   author = {Thomas Manzini and Yao Chong Lim and Yulia Tsvetkov and Alan W Black},
   journal = {arXiv preprint arXiv:1904.04047},
   year = {2019},
+  doi={10.18653/v1/n19-1062}
 }
 
 @InProceedings{mikolov2013distributed,
@@ -200,6 +205,7 @@
   booktitle = {{Proceedings of the 2020 Conference on Fairness, Accountability, and Transparency}},
   pages = {359--368},
   year = {2020},
+  doi={10.1145/3351095.3372837}
 }
 
 @Manual{wijffelsword2vec,
