+++ paper/r-references.bib
@@ -1,15 +1,0 @@
-@Manual{R-base,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2021},
-  url = {https://www.R-project.org/},
-}
-@Manual{R-papaja,
-  author = {Frederik Aust and Marius Barth},
-  title = {{papaja}: {Create} {APA} manuscripts with {R Markdown}},
-  year = {2020},
-  note = {R package version 0.1.0.9997},
-  url = {https://github.com/crsh/papaja},
-}
