### 2021-11-21 #3938 @chainsawriot
I've checked and added all missing DOIs.
