### 2021-11-21 3497887
Fix redact in paper.md
