### 2021-11-21 946ef2a
Fix aff. in paper.md
