### 2021-11-21 #3938 @chainsawriot
I've edited the paper a bit to add the missing DOIs.

May I suggest the following reviewers: pmyteh kbenoit alexhanna
