+++ README.Rmd
@@ -32,10 +32,16 @@ If your goal is to reproduce the analysis in Caliskan et al (2017), please consi
 
 ## Installation
 
-You can install the Github version of sweater with:
+Recommended: You can install the Github version of sweater with:
 
 ``` r
 devtools::install_github("chainsawriot/sweater")
+```
+
+Or the slightly outdated version from CRAN
+
+```r
+install.packages("sweater")
 ```
 
 ## Notation of a query
