### 2021-11-22 83d0502
Remove text2vec dependency, use proxy instead; fix #6
