+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: sweater
 Title: Speedy Word Embedding Association Test and Extras Using R
-Version: 0.1.0
+Version: 0.1.1
 Authors@R: 
     person(given = "Chung-hong",
            family = "Chan",
@@ -22,7 +22,7 @@ Imports:
     purrr,
     quanteda,
     LiblineaR,
-    text2vec,
+    proxy,
     data.table
 Suggests: 
     covr,
