### 2021-11-22 #3938 @Kevin-Mattheus-Moerman
@chainsawriot thanks for this submission and for suggesting potential reviewers. A first step is to confirm whether this work is in scope for JOSS. I have just assigned the `query-scope` label here. This means I have triggered an additional scope review by the editorial board. This is typical for submissions of this size so essentially we will determine whether this work presents with [sufficient scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) to be in scope for JOSS. This scope review should take about 2 weeks.
