### 2021-11-22 59615a0
Initial implementation of RIPA
