+++ tests/testthat/test-ripa.R
@@ -1,0 +1,30 @@
+S2 <- c("janitor", "statistician", "midwife", "bailiff", "auctioneer", 
+"photographer", "geologist", "shoemaker", "athlete", "cashier", 
+"dancer", "housekeeper", "accountant", "physicist", "gardener", 
+"dentist", "weaver", "blacksmith", "psychologist", "supervisor", 
+"mathematician", "surveyor", "tailor", "designer", "economist", 
+"mechanic", "laborer", "postmaster", "broker", "chemist", "librarian", 
+"attendant", "clerical", "musician", "porter", "scientist", "carpenter", 
+"sailor", "instructor", "sheriff", "pilot", "inspector", "mason", 
+"baker", "administrator", "architect", "collector", "operator", 
+"surgeon", "driver", "painter", "conductor", "nurse", "cook", 
+"engineer", "retired", "sales", "lawyer", "clergy", "physician", 
+"farmer", "clerk", "manager", "guard", "artist", "smith", "official", 
+"police", "doctor", "professor", "student", "judge", "teacher", 
+"author", "secretary", "soldier")
+A2 <- c("he", "son", "his", "him", "father", "man", "boy", "himself", 
+"male", "brother", "sons", "fathers", "men", "boys", "males", 
+"brothers", "uncle", "uncles", "nephew", "nephews")
+B2 <- c("she", "daughter", "hers", "her", "mother", "woman", "girl", 
+"herself", "female", "sister", "daughters", "mothers", "women", 
+"girls", "females", "sisters", "aunt", "aunts", "niece", "nieces")
+
+res <- ripa(googlenews, S2, A2, B2)
+
+## write.csv(googlenews, "googlenews_w2v.csv")
+
+## write.csv(data.frame(A2, B2), "word_pairs.csv")
+
+## write.csv(data.frame(S2), "x.csv")
+
+## write.csv(data.frame(S2, p = res$P), "res.csv")
