+++ NAMESPACE
@@ -1,10 +1,13 @@
 # Generated by roxygen2: do not edit by hand
 
 export(calculate_es)
+export(ect)
+export(ect_es)
 export(mac)
 export(mac_es)
 export(nas)
 export(plot_bias)
+export(plot_ect)
 export(query)
 export(read_word2vec)
 export(rnd)
@@ -17,7 +20,9 @@ export(weat_es)
 export(weat_exact)
 export(weat_resampling)
 importFrom(Rcpp,sourceCpp)
+importFrom(graphics,abline)
 importFrom(graphics,dotchart)
+importFrom(graphics,text)
 importFrom(stats,dist)
 importFrom(stats,predict)
 importFrom(stats,sd)
