### 2021-11-23 b38cafb
Fix possible uninitalized variables in the cpp code for exact test
