+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: sweater
 Title: Speedy Word Embedding Association Test and Extras Using R
-Version: 0.1.1
+Version: 0.1.2
 Authors@R: 
     person(given = "Chung-hong",
            family = "Chan",
