### 2021-11-23 c41e64f
Make query, plot_es work for ect
