+++ tests/testthat/test-plot.R
@@ -19,4 +19,5 @@ test_that("weat is not plottable", {
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "semaxis"))), NA)
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "nas"))), NA)
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "rnsb"))), NA)
+    expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "ect"))), NA)
 })
