+++ tests/testthat/test-no_es.R
@@ -51,3 +51,30 @@ test_that("semaxis, garg", {
     expect_true("sweater" %in% class(garg_f1))
     expect_true("semaxis" %in% class(garg_f1))
 })
+
+test_that("ect, garg", {
+    S2 <- c("janitor", "statistician", "midwife", "bailiff", "auctioneer", 
+"photographer", "geologist", "shoemaker", "athlete", "cashier", 
+"dancer", "housekeeper", "accountant", "physicist", "gardener", 
+"dentist", "weaver", "blacksmith", "psychologist", "supervisor", 
+"mathematician", "surveyor", "tailor", "designer", "economist", 
+"mechanic", "laborer", "postmaster", "broker", "chemist", "librarian", 
+"attendant", "clerical", "musician", "porter", "scientist", "carpenter", 
+"sailor", "instructor", "sheriff", "pilot", "inspector", "mason", 
+"baker", "administrator", "architect", "collector", "operator", 
+"surgeon", "driver", "painter", "conductor", "nurse", "cook", 
+"engineer", "retired", "sales", "lawyer", "clergy", "physician", 
+"farmer", "clerk", "manager", "guard", "artist", "smith", "official", 
+"police", "doctor", "professor", "student", "judge", "teacher", 
+"author", "secretary", "soldier")
+    A2 <- c("he", "son", "his", "him", "father", "man", "boy", "himself", 
+"male", "brother", "sons", "fathers", "men", "boys", "males", 
+"brothers", "uncle", "uncles", "nephew", "nephews")
+    B2 <- c("she", "daughter", "hers", "her", "mother", "woman", "girl", 
+"herself", "female", "sister", "daughters", "mothers", "women", 
+"girls", "females", "sisters", "aunt", "aunts", "niece", "nieces"
+)
+    garg_f1 <- ect(googlenews, S2, A2, B2)
+    expect_true("sweater" %in% class(garg_f1))
+    expect_true("ect" %in% class(garg_f1))
+})
