### 2021-11-24 aa7ff5a
Update ect's test case for emitting P
