### 2021-11-24 2cd0af5
Test also `plot_ect`
