### 2021-11-24 3970b4b
Add Embedding Coherence Test ref #8 (#9)
* Initial implementation of ect

* Make query, plot_es work for ect

* Add test cases for ect

* Update the README for ect

* Update ect's test case for emitting P

* Test also `plot_ect`
