+++ tests/testthat/test-query.R
@@ -35,11 +35,13 @@ test_that("guessing and integration with query", {
     expect_equal("weat", .guess(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1, method = "weat"))
     expect_equal("rnd", .guess(w = googlenews, S_words = S1, T_words = T, A_words = A1, B_words = B1, method = "rnd"))
     expect_equal("rnsb", .guess(w = googlenews, S_words = S1, T_words = T, A_words = A1, B_words = B1, method = "rnsb"))
+    expect_equal("ect", .guess(w = googlenews, S_words = S1, T_words = T, A_words = A1, B_words = B1, method = "ect"))
     expect_equal("nas", .guess(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1, method = "nas"))
     expect_true("mac" %in% class(query(w = googlenews, S_words = S1, A_words = A1)))
     expect_true("mac" %in% class(query(w = googlenews, S_words = S1, A_words = A1, method = "mac")))
     expect_true("rnd" %in% class(query(w = googlenews, S_words = S1, A_words = A1, B_words = B1)))
     expect_true("rnd" %in% class(query(w = googlenews, S_words = S1, A_words = A1, B_words = B1, method = "rnd")))
+    expect_true("ect" %in% class(query(w = googlenews, S_words = S1, A_words = A1, B_words = B1, method = "ect")))
     expect_output(query(w = googlenews, S_words = S1, A_words = A1, B_words = B1, verbose = TRUE))
     expect_true("semaxis" %in% class(query(w = googlenews, S_words = S1, A_words = A1, B_words = B1, method = "semaxis")))
     expect_true("weat" %in% class(query(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1)))
@@ -47,4 +49,5 @@ test_that("guessing and integration with query", {
     expect_true("rnd" %in% class(query(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1, method = "rnd")))
     expect_true("rnsb" %in% class(query(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1, method = "rnsb")))
     expect_true("nas" %in% class(query(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1, method = "nas")))
+    expect_true("ect" %in% class(query(w = googlenews, S_words = S1, T_words = T1, A_words = A1, B_words = B1, method = "ect")))
 })
