+++ tests/testthat/test-plot.R
@@ -14,9 +14,11 @@ test_that("weat is not plottable", {
     B1 <- c("female", "woman", "girl", "sister", "she", "her", "hers", "daughter")
     sw <- weat(glove_math, S1, T1, A1, B1)
     expect_error(plot_bias(sw))
+    expect_error(plot_ect(sw))
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words= A1, method = "mac"))), NA)
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "rnd"))), NA)
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "semaxis"))), NA)
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "nas"))), NA)
     expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "rnsb"))), NA)
+    expect_error(save_png(plot_bias(query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = "ect"))), NA)
 })
