### 2021-12-01 fe59414
Implement S3 print and plot methods, fix #10
