### 2021-12-01 0b49486
Fix test case test-plot
