### 2021-12-01 84e4856
Bump version and add reference to Dev et al in DESCRIPTION
