### 2021-12-03 #3938 @kthyng
Hi @chainsawriot! I've removed the query-scope label because the editorial board thought this is in scope. So, we can proceed.

@sbenthall Would you be able to take this on as editor?
