### 2021-12-03 #3938 @kthyng
note for future AEIC: If sbenthall is not able to do this, I currently recommend putting this in a waitlist for an editor in the social sciences/humanities/text-related fields to open up. Choices include chartgerink, gkthiruvathukal, and ajstewartlang.
