### 2021-12-03 #3938 @kthyng
@chainsawriot One note that came up multiple times in feedback from the editorial board is that the "speedy" claim is dubious and that if they were judging based on that claim, this would be deemed not in scope. So, perhaps that should not be an emphasis or perhaps you can elaborate on the claim if you do want to continue in that direction.
