### 2021-12-03 #3938 @chainsawriot
@kthyng Thank you for taking care of the paper.

I can provide a benchmark later to back the speedy claim. If it is not convincing, then I can be creative and change the name of the package to a recursive acronym (sweater: Word Embedding Association Test and Extras Using R).
