### 2021-12-04 #3938 @chainsawriot
@sbenthall Thanks for being the editor of the paper. Previously, I have suggested some potential reviewers. They are listed in the reviewer database and I know they are experts in either text analysis, developing text analytic software, or algorithmic accountability. You might also consider them.

https://github.com/openjournals/joss-reviews/issues/3938#issuecomment-974899645
