+++ paper/benchmark.md
@@ -1,0 +1,195 @@
+
+## Benchmark
+
+This is a version of WEAT written entirely in R.
+
+``` r
+require(purrr)
+```
+
+    ## Loading required package: purrr
+
+``` r
+require(lsa)
+```
+
+    ## Loading required package: lsa
+
+    ## Loading required package: SnowballC
+
+``` r
+take <- function(word, w) {
+    return(as.vector(w[word, , drop = FALSE]))
+}
+
+get_x <- function(w, words) {
+    purrr::map(words, take, w = w)
+}
+
+g <- function(c, A, B, w) {
+    A_emb <- get_x(w, A)
+    B_emb <- get_x(w, B)
+    c_emb <- get_x(w, c)[[1]]
+    a_cos_diff <- mean(purrr::map_dbl(A_emb, ~cosine(., c_emb)))
+    b_cos_diff <- mean(purrr::map_dbl(B_emb, ~cosine(., c_emb)))
+    return(a_cos_diff - b_cos_diff)
+}
+
+.clean <- function(x, w_lab, verbose = FALSE) {
+    new_x <- intersect(x, w_lab)
+    if (length(new_x) < length(x) & verbose) {
+        print("Some word(s) are not available in w.")
+    }
+    return(new_x)
+}
+
+
+r_weat <- function(w, S, T, A, B, verbose = FALSE) {
+    w_lab <- rownames(w)
+    A <- .clean(A, w_lab, verbose = verbose)
+    B <- .clean(B, w_lab, verbose = verbose)
+    S <- .clean(S, w_lab, verbose = verbose)
+    T <- .clean(T, w_lab, verbose = verbose)
+    S_diff <- purrr::map_dbl(S, g, A, B, w)
+    T_diff <- purrr::map_dbl(T, g, A, B, w)
+    ## union_diff <- purrr::map_dbl(union(S, T), g, A, B, w)
+    return((mean(S_diff) - mean(T_diff)) / sd(c(S_diff, T_diff)))
+}
+require(compiler)
+```
+
+    ## Loading required package: compiler
+
+``` r
+r_weat_c <- cmpfun(r_weat)
+```
+
+## The Calikskan et al. example.
+
+``` r
+require(sweater)
+```
+
+    ## Loading required package: sweater
+
+    ## 
+    ## Attaching package: 'sweater'
+
+    ## The following object is masked from 'package:lsa':
+    ## 
+    ##     query
+
+``` r
+S2 <- c("math", "algebra", "geometry", "calculus", "equations",
+        "computation", "numbers", "addition")
+T2 <- c("poetry", "art", "dance", "literature", "novel", "symphony",
+        "drama", "sculpture")
+A2 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
+B2 <- c("female", "woman", "girl", "sister", "she", "her", "hers",
+        "daughter")
+r_weat(glove_math, S2, T2, A2, B2)
+```
+
+    ## [1] 1.055015
+
+``` r
+r_weat_c(glove_math, S2, T2, A2, B2)
+```
+
+    ## [1] 1.055015
+
+The same implementation in C++ from `sweater`
+
+``` r
+calculate_es(query(glove_math, S2, T2, A2, B2))
+```
+
+    ## [1] 1.055015
+
+``` r
+cpp_weat <- function(w, S, T, A, B) {
+    calculate_es(query(w, S, T, A, B))
+}
+```
+
+The C++ implementation in `sweater` is \>10x faster. Byte-code
+compilation (`r_weat_c`) can bring about almost no little improvement.
+
+``` r
+require(bench)
+```
+
+    ## Loading required package: bench
+
+``` r
+benchmark_res <- bench::mark(
[…]
