### 2021-12-06 608e543
Update paper.rmd to reflect the added ECT
