+++ paper/paper.bib
@@ -225,3 +225,12 @@
   isbn	       = 9781461468684,
   publisher    = {Springer New York}
 }
+
+@inproceedings{dev2019attenuating,
+  title={Attenuating bias in word vectors},
+  author={Dev, Sunipa and Phillips, Jeff},
+  booktitle={The 22nd International Conference on Artificial Intelligence and Statistics},
+  pages={879--887},
+  year={2019},
+  organization={PMLR}
+}
