### 2021-12-07 7aed490
Fix the S3 test snapshot
