+++ tests/testthat/_snaps/s3.md
@@ -3,13 +3,11 @@
     Code
       sw
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  weat 
       Effect size:  1.055015 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <weat_resampling()>: Conduct statistical test
@@ -19,13 +17,11 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, method = "mac")
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  mac 
       Effect size:  0.09579005 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
@@ -35,13 +31,11 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  rnd 
       Effect size:  -1.147534 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
@@ -51,12 +45,10 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  semaxis 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <plot()>: Plot the bias of each individual word
 
@@ -65,12 +57,10 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  nas 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <plot()>: Plot the bias of each individual word
 
@@ -79,13 +69,11 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  rnsb 
       Effect size:  0.01772161 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
@@ -95,13 +83,11 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
-      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  ect 
       Effect size:  0.9761905 
     Message <cliMessage>
-      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
