### 2021-12-07 9c0406b
Integrate benchmark into paper and README
