+++ paper/paper.rmd
@@ -30,7 +30,9 @@ set.seed(46709394)
 
 The goal of this R package is to detect (implicit) biases in word embeddings. The importance of detecting biases in word embeddings is twofold. First, pretrained, biased word embeddings deployed in real-life machine learning systems can pose fairness concerns [@packer2018text;@boyarskaya2020overcoming]. Second, biases in word embeddings reflect the biases in the original training material. Social scientists, communication researchers included, have exploited these methods to quantify (implicit) media biases by extracting biases from word embeddings locally trained on large text corpora [e.g. @kroon2020guilty;@knoche2019identifying;@sales2019media]. Biases in word embedding can be understood through the implicit social cognition model of media priming [@arendt:2013:DDM]. In this model, implicit stereotypes are defined as the "strength of the automatic association between a group concept (e.g., minority group) and an attribute (e.g., criminal)." [@arendt:2013:DDM, p. 832] All of these bias detection methods are based on the strength of association between a concept (or a target) and an attribute in embedding spaces.
 
-Previously, the software of these methods is only scatteredly available as the addendum of the original papers and was implemented in different languages (Java, Python, etc.). `sweater` provides several of these bias detection methods in one unified package with a consistent R interface [@rcore]. Also, some provided methods are implemented in C++ for speed and interfaced to R using the `Rcpp` package [@eddelbuettel:2013:SRC].
+Previously, the software of these methods is only scatteredly available as the addendum of the original papers and was implemented in different languages (Java, Python, etc.). `sweater` provides several of these bias detection methods in one unified package with a consistent R interface [@rcore]. Also, some provided methods are implemented in C++ for speed and interfaced to R using the `Rcpp` package [@eddelbuettel:2013:SRC] [^BENCH].
+
+[^BENCH]: Compared with a pure R implementation, the C++ implementation of Word Embedding Association Test in `sweater` is at least 7 times faster. See the benchmark [here](https://github.com/chainsawriot/sweater/blob/master/paper/benchmark.md).
 
 In the usage section below, we demonstrated how the package can be used to detect biases and reproduce some published findings.
 
