### 2022-01-06 #4036 @whedon
[REVIEW]: sweater: Speedy Word Embedding Association Test and Extras Using R
**Submitting author:** @chainsawriot (<a href="http://orcid.org/0000-0002-6232-7530">Chung-hong Chan</a>)
**Repository:** https://github.com/chainsawriot/sweater
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.4
