+++ paper/ica.tex
@@ -1,7 +1,7 @@
 \PassOptionsToPackage{unicode=true}{hyperref} % options for packages loaded elsewhere
 \PassOptionsToPackage{hyphens}{url}
 %
-\documentclass[english,man,mask]{apa6}
+\documentclass[english,man]{apa6}
 \usepackage{lmodern}
 \usepackage{amssymb,amsmath}
 \usepackage{ifxetex,ifluatex}
@@ -174,7 +174,7 @@ Correspondence concerning this article should be addressed to Chung-hong Chan, A
 \affiliation{\vspace{0.5cm}\textsuperscript{1} Mannheimer Zentrum für Europäische Sozialforschung}
 
 \abstract{
-This paper shows how the R package \{redacted\} can be used to detect biases in word embeddings. The package provides highly optimized functions to calculate the following bias metrics: mean average cosine similarity, relative norm distance, SemAxis, normalized association score, relative negative sentiment bias, and word embedding association test. Using two public available word embeddings trained on media content, this paper demonstrates how \{redacted\} can be used to study implicit gender stereotypes.
+This paper shows how the R package sweater can be used to detect biases in word embeddings. The package provides highly optimized functions to calculate the following bias metrics: mean average cosine similarity, relative norm distance, SemAxis, normalized association score, relative negative sentiment bias, embedding coherence test and word embedding association test. Using two public available word embeddings trained on media content, this paper demonstrates how sweater can be used to study implicit gender stereotypes.
 }
 
 
@@ -185,9 +185,9 @@ This paper shows how the R package \{redacted\} can be used to detect biases in 
 \hypertarget{statement-of-need}{%
 \section{Statement of need}\label{statement-of-need}}
 
-The goal of the R package \{redacted\} is to detect (implicit) biases in word embeddings. The importance of detecting biases in word embeddings is twofold. First, pretrained, biased word embeddings deployed in real-life machine learning systems can pose fairness concerns (Boyarskaya, Olteanu, \& Crawford, 2020; Packer, Mitchell, Guajardo-Céspedes, \& Halpern, 2018). Second, biases in word embeddings reflect the biases in the original training material. Social scientists, communication researchers included, have exploited these methods to quantify (implicit) media biases by extracting biases from word embeddings locally trained on large text corpora (e.g.~Kroon, Trilling, \& Raats, 2020; Knoche, Popović, Lemmerich, \& Strohmaier, 2019; Sales, Balby, \& Veloso, 2019). Biases in word embedding can be understood through the implicit social cognition model of media priming (Arendt, 2013). In this model, implicit stereotypes are defined as the \enquote{strength of the automatic association between a group concept (e.g., minority group) and an attribute (e.g., criminal).} (Arendt, 2013, p. 832) All of these bias detection methods are based on the strength of association between a concept (or a target) and an attribute in embedding spaces.
[…]
