### 2022-01-20 b57b350
Fix the failing test cases in r-devel
