### 2022-01-26 chainsawriot/sweater#11 @cmaimone
installation instructions
Suggesting people use `remotes::install_github("chainsawriot/sweater")` instead of `devtools::install_github("chainsawriot/sweater")` will probably result in fewer issues, as remotes is easier for users to install successfully.

Would also consider rewording: "Or the slightly outdated version from CRAN" -- if you're going to offer via CRAN, that should be kept reasonably up to date, and the github version should be considered the developmental one.

re: https://github.com/openjournals/joss-reviews/issues/4036
