### 2022-01-26 48daf78
suggested clarifications
on acceptable word embedding formats

re: https://github.com/openjournals/joss-reviews/issues/4036
