### 2022-01-26 chainsawriot/sweater#13 @cmaimone
Paper: questions on Query section
For this:

`sweater` uses the concept of *query* [@badilla2020wefe] to study the biases in $w$. A query contains two or more sets of seed words with at least one set of *target words* and one set of *attribute words*. `sweater` uses the $\mathcal{S}\mathcal{T}\mathcal{A}\mathcal{B}$ notation from @brunet2019understanding to form a query.

* Need: concept of a *query* (missing a)
* Why is STAB in mathematical notation?
* Are target words and attribute words types of seed words?  I think so, but that could be clearer

I would also find a little more info on target and attribute sets helpful.  When you're supposed to supply two different sets, what is each supposed to be?  What should be in S and what in T?  I appreciate the references, and realize this may be complicated.  Some type of brief summary here would help though.  For example, for A and B, it seems each should be a set of words relevant to a group?  Or the endpoints of a scale?

When you say target words shouldn't have bias, does that mean they are the words you're testing for bias?
