### 2022-01-27 chainsawriot/sweater#13 @cmaimone
This is re: https://github.com/openjournals/joss-reviews/issues/4036
