### 2022-01-27 chainsawriot/sweater#14 @cmaimone
speed claims?
I'm not completely clear on what the speed claims are here.  What is sweater faster than exactly?

I think benchmark.md is showing that implementing one method with rcpp is faster than implementing the same method in pure R or R running C code?

That doesn't seem sufficient for a speed claim in general, especially with that in the package name.

It's very possible I'm missing something though.

I'll note that, other than what's implied by the package name, the paper doesn't make specific speed/performance claims.  Dropping the last sentence of the second paragraph of the paper would, I think, remove all reference to speed.  If you don't want to go into full comparisons and benchmarking, it may just be enough to talk about how long sweater takes to do common things, and refrain from saying whether that's fast or slow.   Having that information -- expected run times -- is very useful just by itself.

re: https://github.com/openjournals/joss-reviews/issues/4036
