### 2022-01-27 chainsawriot/sweater#14 @cmaimone
Documentation (readme.md) says: "or are speedy but accurate approximation of the original implementation proposed by Caliskan et al (2017)"

How are you determining that the approximations are accurate?  Can you quantify the accuracy in any way?
