### 2022-01-27 chainsawriot/sweater#15 @cmaimone
paper citations
Looks like the paper cites the R word2vec package, but not the original dataset.  Should original dataset be cited?  Same for small_reddit, glove_math?

Perhaps list them in the paper as sample word embeddings that are included in the package?

re: https://github.com/openjournals/joss-reviews/issues/4036
