### 2022-01-27 chainsawriot/sweater#16 @cmaimone
paper: other R packages?
Are there any other R packages available for bias in word embeddings, or more generally for bias in text corpora?  If not, you can state that.  If there are, would be good to reference.

The statement that sweater brings together methods that were only reported in papers' supplemental materials is clear -- I get the utility of the package.  I just don't know if there's anything else out there or not.

re: https://github.com/openjournals/joss-reviews/issues/4036
