### 2022-01-27 chainsawriot/sweater#17 @cmaimone
paper suggestion
Suggestion: intro to the paper would be stronger/clearer with an example of what implicit word biases are -- that neutral seeming words can be more associated with one gender than the other, or one ethnic group over another.  Which can lead to...

Just a few sentences.

Not absolutely needed, but it jumps pretty quickly into social science theory without a generally accessible example of what it is.

re: https://github.com/openjournals/joss-reviews/issues/4036
