### 2022-01-27 chainsawriot/sweater#18 @cmaimone
community guidelines
Add something about how people seeking support with the software should get it -- where can/should people ask questions?  It's not very clear from the current contributing section.

re: https://github.com/openjournals/joss-reviews/issues/4036
