### 2022-01-27 chainsawriot/sweater#19 @cmaimone
documentation: what is the preferred workflow?
> It is recommended to use the function query() to make a query and calculate_es() to calculate the effect size. You can also use the functions listed below.

I'm confused why there are both general functions and specific ones -- seems like you should either specify the method as an argument to query() or have to use a query function specific to each method -- not both options.

Are functions mac(), rnd(), nas() etc functions that create a query?  If so, it would be great to get "query" in their names to make that clear.

re: https://github.com/openjournals/joss-reviews/issues/4036
