### 2022-01-27 chainsawriot/sweater#20 @cmaimone
documentation: first example
The first example uses googlenews without introducing it first -- not even that it has word embeddings in it.

Output from the first example isn't explained.  I found it confusing.

```
#> Effect size:  0.1375856
#>
#> ── Functions ─────────────────────────────────
#> • <calculate_es()>: Calculate effect size
#> • <plot()>: Plot the bias of each individual word
```

Why are the function names in <>?  Why are these listed?  Are these next steps that I should take?  Do I supply the output of the query function to them?

What is mac_neg$P?

re: https://github.com/openjournals/joss-reviews/issues/4036
