### 2022-01-27 chainsawriot/sweater#21 @cmaimone
"guess" method
I found it confusing in the documentation how the method was being determined.  The examples are listed for specific methods, but the query code doesn't specify the method.  I see that the default is to guess, which I assume it does based on the combination of STAB inputs provided?  But if the examples are supposed to be of a specific method, it would probably be clearer to show code that invokes that method specifically -- what if you change "guess" behavior in the future?

re: https://github.com/openjournals/joss-reviews/issues/4036
