### 2022-01-27 chainsawriot/sweater#22 @cmaimone
documentation: citation links
All of the citations to papers in the documentation are great!  They'd be even more useful if you linked them, so people could click on the reference to go to it.

re: https://github.com/openjournals/joss-reviews/issues/4036
