### 2022-01-27 chainsawriot/sweater#23 @cmaimone
Range of effect sizes?
It would be really helpful if the documentation pages for the different methods (or the effect size function pages?) included information on the scale/range/directional of the effect size output.  For example, are the values between 0 and 1?  What is an example of a lot of bias vs. no bias?

re: https://github.com/openjournals/joss-reviews/issues/4036
