### 2022-01-27 chainsawriot/sweater#25 @cmaimone
code structure question
Are effect sizes expensive to compute?  Looks like a call to query() will compute the effect size automatically and print it out, so my guess is no?  Is there a reason not to store the computed effect size in the sweater objects (so it can be referenced later)?  With the effect size functions, I'd have to store the effect size for a query in a separate variable from the query object, which makes it easy to get things messed up in the code -- would be nice to have it as part of the result object.

Could still have es functions for convenience of accessing that component of the object if you wanted.

re: https://github.com/openjournals/joss-reviews/issues/4036
