### 2022-01-27 chainsawriot/sweater#26 @cmaimone
plotting error
```
> S4 <- c("math", "algebra", "geometry", "calculus", "equations", "computation", "numbers", "addition")
> T4 <- c("poetry", "art", "dance", "literature", "novel", "symphony", "drama", "sculpture")
> A4 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
> B4 <- c("female", "woman", "girl", "sister", "she", "her", "hers", "daughter")
> sw <- query(glove_math, S4, T4, A4, B4)
>
> names(sw)
[1] "S_diff"  "T_diff"  "S_words" "T_words" "A_words" "B_words"
> class(sw)
[1] "sweater" "weat"    "list"
> plot(sw)
Error in plot_bias(x) : No P slot in the input object x.FALSE
```

^ I ran the example, then tried to plot.  Should this be giving this error?

re: https://github.com/openjournals/joss-reviews/issues/4036
