### 2022-01-27 chainsawriot/sweater#27 @cmaimone
names for vectors in output object?
Can the `S_diff` and `T_diff` components of the output for the method have names?  I think each value corresponds to an input term, yes?  Would be more useful as named vectors.

```
> sw$S_diff
[1]  0.003158583  0.003242220  0.001271607  0.031652155  0.003074379  0.016247332  0.035000510 -0.010817083
> sw$T_diff
[1] -0.0265718087  0.0054876842 -0.0523231481 -0.0117847993 -0.0369267966  0.0224587349 -0.0167662057
[8]  0.0003334358
```

re: https://github.com/openjournals/joss-reviews/issues/4036
