### 2022-01-27 chainsawriot/sweater#26 @chainsawriot
@cmaimone  Thanks for raising this.

`weat` S3 object is not plottable. `plot` is not an available S3 method (which is still a function) for this object and it is reflected by printing the object (cf. `rnd`).

``` r
require(sweater)
#> Loading required package: sweater

S4 <- c("math", "algebra", "geometry", "calculus", "equations", "computation", "numbers", "addition")
T4 <- c("poetry", "art", "dance", "literature", "novel", "symphony", "drama", "sculpture")
A4 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
B4 <- c("female", "woman", "girl", "sister", "she", "her", "hers", "daughter")
sw <- query(glove_math, S4, T4, A4, B4)
sw
#>
#> ── sweater object ──────────────────────────────────────────────────────────────
#> Test type:  weat
#> Effect size:  1.055015
#>
#> ── Functions ───────────────────────────────────────────────────────────────────
#> • <calculate_es()>: Calculate effect size
#> • <weat_resampling()>: Conduct statistical test

sw2 <- query(glove_math, S_words = S4, A_words = A4, B_words = B4)
sw2
#>
#> ── sweater object ──────────────────────────────────────────────────────────────
#> Test type:  rnd
#> Effect size:  -1.147534
#>
#> ── Functions ───────────────────────────────────────────────────────────────────
#> • <calculate_es()>: Calculate effect size
#> • <plot()>: Plot the bias of each individual word
```

But the error message can be better, e.g. `object of type 'weat' is not plottable.` [^JOKE] I will amend the error message to something more useful.

[^JOKE]: OK, this is an in-joke. `object of type 'closure' is not subsettable`
