### 2022-01-27 9600f76
Make `S_diff` and `T_diff` named vectors, fix #27
