### 2022-01-27 8bcc18b
Update testthat and S3 snapshot test cases
