+++ tests/testthat/_snaps/s3.md
@@ -3,11 +3,13 @@
     Code
       sw
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  weat 
       Effect size:  1.055015 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <weat_resampling()>: Conduct statistical test
@@ -17,11 +19,13 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, method = "mac")
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  mac 
       Effect size:  0.09579005 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
@@ -31,11 +35,13 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  rnd 
       Effect size:  -1.147534 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
@@ -45,10 +51,12 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  semaxis 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <plot()>: Plot the bias of each individual word
 
@@ -57,10 +65,12 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  nas 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <plot()>: Plot the bias of each individual word
 
@@ -69,11 +79,13 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  rnsb 
       Effect size:  0.01772161 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
@@ -83,11 +95,13 @@
     Code
       query(glove_math, S_words = S1, A_words = A1, B_words = B1, method = test_types)
     Message <cliMessage>
+      
       -- sweater object --------------------------------------------------------------
     Output
       Test type:  ect 
       Effect size:  0.9761905 
     Message <cliMessage>
+      
       -- Functions -------------------------------------------------------------------
       * <calculate_es()>: Calculate effect size
       * <plot()>: Plot the bias of each individual word
