### 2022-01-27 55b2c07
Make `S_diff` and `T_diff` named vectors, fix #27 (#28)
