### 2022-01-27 31bf79a
Make `plot` and `plot_bias` emit more useful error message, fix #26
when `x` is of the type `weat`.
