### 2022-01-27 1c4d6a1
Make `plot` and `plot_bias` emit more useful error message, fix #26 (#29)
when `x` is of the type `weat`.
