### 2022-01-27 a34f3dc
make print method use `.fun` style instead. ref #20
