### 2022-01-27 68fa64a
update the first example; ref #20
