+++ README.Rmd
@@ -72,7 +72,9 @@ It is recommended to use the function `query()` to make a query and `calculate_e
 
 ## Example: Mean Average Cosine Similarity
 
-The simplest form of bias detection is Mean Average Cosine Similarity (Mazini et al. 2019). The same method is used also in Kroon et al. (2020).
+The simplest form of bias detection is Mean Average Cosine Similarity (Mazini et al. 2019). The same method is used also in Kroon et al. (2020). `googlenews` is a subset of [the pretrained word2vec word embeddings provided by Google](https://code.google.com/archive/p/word2vec/).
+
+By default, the `query()` function guesses the method you want to use based on the combination of target words and attribute words provided. You can also make this explicit by specifying the `method` argument.
 
 ```{r, eval = FALSE}
 require(sweater)
@@ -98,9 +100,13 @@ A1 <- c("he", "son", "his", "him", "father", "man", "boy", "himself",
 "male", "brother", "sons", "fathers", "men", "boys", "males", 
 "brothers", "uncle", "uncles", "nephew", "nephews")
 
+## The same as:
+## mac_neg <- query(googlenews, S_words = S1, A_words = A1, method = "mac")
 mac_neg <- query(googlenews, S_words = S1, A_words = A1)
 mac_neg
 ```
+
+The returned object is an S3 object. Please refer to the help file of the method for the definition of all slots (in this case: `?mac`). For example, the magnitude of bias for each word is available in the `P` slot. 
 
 ```{r mac_neg2}
 sort(mac_neg$P)
