### 2022-01-27 chainsawriot/sweater#26 @chainsawriot
@cmaimone I think this error message is more useful.

``` r
require(sweater)
#> Loading required package: sweater

S4 <- c("math", "algebra", "geometry", "calculus", "equations", "computation", "numbers", "addition")
T4 <- c("poetry", "art", "dance", "literature", "novel", "symphony", "drama", "sculpture")
A4 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
B4 <- c("female", "woman", "girl", "sister", "she", "her", "hers", "daughter")
sw <- query(glove_math, S4, T4, A4, B4)
sw
#>
#> ── sweater object ──────────────────────────────────────────────────────────────
#> Test type:  weat
#> Effect size:  1.055015
#>
#> ── Functions ───────────────────────────────────────────────────────────────────
#> • <calculate_es()>: Calculate effect size
#> • <weat_resampling()>: Conduct statistical test
plot(sw)
#> Error: Plotting the result of this test type (weat) is not supported.
plot_bias(sw)
#> Error: Plotting the result of this test type (weat) is not supported.
```

Also it is now explicitly documented in the man page.

https://github.com/chainsawriot/sweater/blob/68fa64af1d7a8bf547da7df5e6681365b57787a4/man/plot_bias.Rd#L21
