### 2022-01-28 chainsawriot/sweater#16 @chainsawriot
@cmaimone As far as I know, there is only one R package available aimed at replicating the analysis in Caliskan et al. (https://github.com/conjugateprior/cbn , i.e. only `weat`) Other than that, I don't think there are R packages providing other methods.

The Python package [wefe](https://github.com/dccuchile/wefe) is kind of similar to this package.

I can mention these in the paper.
