+++ paper/paper.rmd
@@ -14,8 +14,8 @@ affiliations:
  - name: Mannheimer Zentrum für Europäische Sozialforschung, Universität Mannheim
    index: 1
 citation_author: Chan.
-date: 20 November 2021
-year: 2021
+date: 28 January 2022
+year: 2022
 bibliography: paper.bib
 output: rticles::joss_article
 csl: apa.csl
@@ -34,9 +34,13 @@ Previously, the software of these methods is only scatteredly available as the a
 
 [^BENCH]: Compared with a pure R implementation, the C++ implementation of Word Embedding Association Test in `sweater` is at least 7 times faster. See the benchmark [here](https://github.com/chainsawriot/sweater/blob/master/paper/benchmark.md).
 
-In the usage section below, we demonstrated how the package can be used to detect biases and reproduce some published findings.
+# Related work
+
+The R package `cbn` (https://github.com/conjugateprior/cbn) by Will Lowe provides tools for replicating the study by @caliskan:2017:S. The Python package `wefe` [@badilla2020wefe] provides several methods for bias evaluation in a unified (Python) interface.
 
 # Usage
+
+In this section, I demonstrate how the package can be used to detect biases and reproduce some published findings.
 
 ## Word Embeddings
 
