### 2022-01-28 chainsawriot/sweater#19 @chainsawriot
@cmaimone As of now, the preferred workflow is certainly `query`. The monolithic `query` interface is unfortunately an afterthought (4c0751c2b235265e807cf87e0cb813a170c914d0) and the package started with individual functions for each method. I know that there has been legacy code in some other researchers' analytic scripts using those individual functions. For backward compatibility, I need to keep those functions — their names at least — as they were (also, it's too late now to rename those functions with some tidyverse-style verb-based function names, similar to your "get `query` in their names").

Maybe a more proper approach is to make those individual functions less prominent in the README and state that they are here for backward compatibility.
