### 2022-01-28 c5bfe15
suggested clarifications (#12)
on acceptable word embedding formats

re: https://github.com/openjournals/joss-reviews/issues/4036
