### 2022-01-28 chainsawriot/sweater#19 @cmaimone
Downplaying them in the readme seems like a good strategy -- maybe even mark them as depreciated and plan to take them out in the future (or, if you prefer separate functions, change the names)?  Your package is pretty young to feel like you're locked in forever, but I do appreciate the care for existing users.
