+++ paper/benchmark.md
@@ -133,9 +133,9 @@ benchmark_res
     ## # A tibble: 3 × 6
     ##   expression                             min median `itr/sec` mem_alloc `gc/sec`
     ##   <bch:expr>                           <dbl>  <dbl>     <dbl>     <dbl>    <dbl>
-    ## 1 r_weat(glove_math, S2, T2, A2, B2)    12.0   11.4      1.04      2.35     1   
-    ## 2 r_weat_c(glove_math, S2, T2, A2, B2)  13.0   11.8      1         2.35     1.01
-    ## 3 cpp_weat(glove_math, S2, T2, A2, B2)   1      1       11.1       1        2.06
+    ## 1 r_weat(glove_math, S2, T2, A2, B2)    1.30   1.33      1.01      2.35     7.63
+    ## 2 r_weat_c(glove_math, S2, T2, A2, B2)  1.30   1.35      1         2.35     8.98
+    ## 3 cpp_weat(glove_math, S2, T2, A2, B2)  1      1         1.24      1        1
 
 ### Random benchmark
 
@@ -173,8 +173,6 @@ res <- map(stab_length, r_bench)
     ## Warning: Some expressions had a GC in every iteration; so filtering is disabled.
     
     ## Warning: Some expressions had a GC in every iteration; so filtering is disabled.
-    
-    ## Warning: Some expressions had a GC in every iteration; so filtering is disabled.
 
 ``` r
 res %>% map_dfr(~.[1,3]) %>% dplyr::mutate(stab_length = stab_length)
@@ -183,13 +181,129 @@ res %>% map_dfr(~.[1,3]) %>% dplyr::mutate(stab_length = stab_length)
     ## # A tibble: 10 × 2
     ##    median stab_length
     ##     <dbl>       <dbl>
-    ##  1  10.2           10
-    ##  2  10.2           20
-    ##  3   8.04          30
-    ##  4   9.46          40
-    ##  5  13.3           50
-    ##  6  11.9           60
-    ##  7   8.98          70
-    ##  8   8.20          80
-    ##  9   7.98          90
-    ## 10  10.1          100
+    ##  1   1.09          10
+    ##  2   1.01          20
+    ##  3   2.44          30
+    ##  4   1.24          40
+    ##  5   1.09          50
+    ##  6   1.02          60
+    ##  7   1.02          70
+    ##  8   1             80
+    ##  9   1             90
+    ## 10   1.01         100
+
+### versus WEFE
+
+The hidden gem of this package is the function `read_word2vec`. It is so
+flexible and yet speedy. In the following example, we are going to
+compare the typical task of the bias detection pipeline:
+
+1.  read a pretrained word embedding file (In this case GloVE, 5.3 GB)
+2.  do WEAT
+
+`read_word2vec` is based on the C++ based function `data.table::fread`
+and it can read almost all formats (word2vec, glove, fastText, etc). The
+entire workflow can be finished in less than a minute.
+
+``` bash
+time Rscript bench.R
+```
+
+    ## Loading required package: sweater
+    ## 
+    ## ── sweater object ──────────────────────────────────────────────────────────────
+    ## Test type:  weat 
+    ## Effect size:  1.055015 
+    ## 
+    ## ── Functions ───────────────────────────────────────────────────────────────────
+    ## • <calculate_es()>: Calculate effect size
[…]
