### 2022-01-28 chainsawriot/sweater#14 @chainsawriot
@cmaimone Regarding the speed claims, in openjournals/joss-reviews#3938 @kthyng also perceived the speed claims of this package as dubious. And thus I added a benchmark.

I still believe this package provides heavily optimized solutions, especially compared to the existing solutions out there. The comparison with native and compiled R looks artificial. How about the comparison between the equivalent Python's offering? I've updated the benchmark (in the `paper` branch) and compare sweater with wefe. sweater is >10x faster (analyzing 5GB of data, 1 min vs 12.5 mins), thanks to the heavily optimized `read_word2vec`.

https://github.com/chainsawriot/sweater/blob/paper/paper/benchmark.md

(I am also trying to add the [original Java code](https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/DX4VWP/ENPK48&version=2.0) by Caliskan. But I need some time.)

Having said that, I know that all benchmarks can lead to controversies. The "speedy" part of the name was chosen to indicate that the package has done some optimization in a low-level language (C++). I don't know how much optimization should be done to qualify this package as "speedy". And as things turn out, the "speedy" name is a mistake. I don't mind dropping all speed claims. I also don't mind renaming the long form of the package name too. I could rename the package to a recursive acronym, eg. "sweater: word embedding association test & extras".

Regarding the accuracy, the original exact test in Caliskan is not practical in real life applications, e.g. when the word set length is large than 10. It took an unrealistic amount of time to calculate even after C++ optimization (I can never finish one, but my estimation is a few day). I will add a vignette to explain how accurate it is.

TODOS

- [x] accuracy vignette
- [x] java benchmark
