### 2022-01-28 chainsawriot/sweater#19 @chainsawriot
TODO

- [x] downplay those functions in the README
- [x] improve documentation regarding the relationship between `query` and those functions
