### 2022-01-28 chainsawriot/sweater#14 @cmaimone
It's up to you which direction you go (more benchmarking vs. name change).  I don't have strong opinions here (on direction or what's sufficient) -- mostly just going through the JOSS checklist, and addressing the performance claims point.  Maybe @kbenoit (other reviewer) has some thoughts on this point?
