+++ README.Rmd
@@ -55,26 +55,26 @@ All tests depend on two types of words. The first type, namely, `S_words` and `T
 
 The second type, namely `A_words` and `B_words`, is *attribute words* (or *group words* in Garg et al). These are words with known properties in relation to the bias that one is studying. For example, Caliskan et al. (2017) used gender-related words such as "male", "man", "boy", "brother", "he", "him", "his", "son" to study gender bias. These words qualify as attribute words because we know they are related to a certain gender.
 
-It is recommended to use the function `query()` to make a query and `calculate_es()` to calculate the effect size. You can also use the functions listed below.
+It is recommended to use the function `query()` to make a query and `calculate_es()` to calculate the effect size.
 
 ## Available methods
 
-| Target words     | Attribute words  | Method                                                      | functions                                          |
-|------------------|------------------|-------------------------------------------------------------|----------------------------------------------------|
-| S_words          | A_words          | Mean Average Cosine Similarity (Mazini et al. 2019)         | mac(), mac_es()                                    |
-| S_words          | A_words, B_words | Relative Norm Distance (Garg et al. 2018)                   | rnd(), rnd_es()                                    |
-| S_words          | A_words, B_words | Relative Negative Sentiment Bias (Sweeney & Najafian. 2019) | rnsb(), rnsb_es()                                  |
-| S_words          | A_words, B_words | Embedding Coherence Test (Dev & Phillips. 2019)             | ect(), ect_es(), plot_ect()                        |
-| S_words          | A_words, B_words | SemAxis (An et al. 2018)                                    | semaxis()                                          |
-| S_words          | A_words, B_words | Normalized Association Score (Caliskan et al. 2017)         | nas()                                              |
-| S_words, T_words | A_words, B_words | Word Embedding Association Test (Caliskan et al. 2017)      | weat(), weat_es(), weat_resampling(), weat_exact() |
-| S_words, T_words | A_words, B_words | Word Embeddings Fairness Evaluation (Badilla et al. 2020)   | To be implemented                                  |
+| Target words     | Attribute words  | Method                                                      | `method` argument | Suggested by `query`? | legacy functions [^legacy]                         |
+|------------------|------------------|-------------------------------------------------------------|-------------------|-----------------------|----------------------------------------------------|
+| S_words          | A_words          | Mean Average Cosine Similarity (Mazini et al. 2019)         | "mac"             | yes                   | mac(), mac_es()                                    |
+| S_words          | A_words, B_words | Relative Norm Distance (Garg et al. 2018)                   | "rnd"             | yes                   | rnd(), rnd_es()                                    |
+| S_words          | A_words, B_words | Relative Negative Sentiment Bias (Sweeney & Najafian. 2019) | "rnsb"            | no                    | rnsb(), rnsb_es()                                  |
+| S_words          | A_words, B_words | Embedding Coherence Test (Dev & Phillips. 2019)             | "ect"             | no                    | ect(), ect_es(), plot_ect()                        |
+| S_words          | A_words, B_words | SemAxis (An et al. 2018)                                    | "semaxis"         | no                    | semaxis()                                          |
+| S_words          | A_words, B_words | Normalized Association Score (Caliskan et al. 2017)         | "nas"             | no                    | nas()                                              |
+| S_words, T_words | A_words, B_words | Word Embedding Association Test (Caliskan et al. 2017)      | "weat"            | yes                   | weat(), weat_es(), weat_resampling(), weat_exact() |
[…]
