### 2022-01-29 chainsawriot/sweater#21 @chainsawriot
644d71d makes it explicit about what the "guess" method does.

TODO
- [x] Make it also explicit in the help files
