### 2022-01-29 aeb6124
Update the paper, readme and benchmark with new printing style
