+++ paper/paper.rmd
@@ -76,8 +76,15 @@ Table: All methods supported by sweater
 
 Relative Norm Distance (RND) [@garg:2018:W] is calculated with two sets of attribute words. The following analysis reproduces the calculation of "women bias" values in @garg:2018:W. The publicly available word2vec word embeddings trained on the Google News corpus is used [@mikolov2013distributed]. Words such as "nurse", "midwife" and "librarian" are more associated with female, as indicated by the positive relative norm distance (Figure 1).
 
-```{r rnd, echo = TRUE, fig.height = 11, fig.cap = "Bias of words in the target wordset according to relative norm distance"}
+```{r echo = FALSE, message = FALSE}
+devtools::load_all()
+```
+
+```{r eval = FALSE}
 library(sweater)
+```
+
+```{r rnd, echo = TRUE, fig.height = 11, fig.cap = "Bias of words in the target wordset according to relative norm distance"}
 data(googlenews)
 S1 <- c("janitor", "statistician", "midwife", "bailiff", "auctioneer",
        "photographer", "geologist", "shoemaker", "athlete", "cashier",
