### 2022-01-29 34d0bef
Make print method use `.fun` style instead. ref #20 (#31)
* make print method use `.fun` style instead. ref #20

* Update snapshot to use the new printing style

* Update the paper, readme and benchmark with new printing style

* Recompile C++
