### 2022-01-29 chainsawriot/sweater#20 @chainsawriot
@cmaimone I have rewritten the first example and it is now clearer (I think). Concerning the angle brackets, it is a bit confusing. I just found that the latest version of `cli` has a new style called `.fun` specifically for printing functions (at least that's how the RStudio team suggested).

I see some packages are using `.fun` now, e.g. `crossjoin` here.

![image](https://user-images.githubusercontent.com/637269/151660249-35ad6e22-e6b5-48dd-bbdc-8826bb887f81.png)

 I hope it is better now, although it still has a pair of surrounding backticks.

``` r
require(sweater)
#> Loading required package: sweater
S4 <- c("math", "algebra", "geometry", "calculus", "equations", "computation", "numbers", "addition")
A4 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
query(glove_math, S_words = S4, A_words = A4)
#>
#> ── sweater object ──────────────────────────────────────────────────────────────
#> Test type:  mac
#> Effect size:  0.09579005
#>
#> ── Functions ───────────────────────────────────────────────────────────────────
#> • `calculate_es()`: Calculate effect size
#> • `plot()`: Plot the bias of each individual word
```
