### 2022-01-29 chainsawriot/sweater#25 @chainsawriot
@cmaimone effect sizes are not expensive to compute and for many methods there is only one effect size. But the reason I don't want to put it inside the S3 is that for WEAT there are several ways to calculate ES. It would create the problem of which ES to put inside the S3.

``` r
require(sweater)
#> Loading required package: sweater

S4 <- c("math", "algebra", "geometry", "calculus", "equations", "computation", "numbers", "addition")
T4 <- c("poetry", "art", "dance", "literature", "novel", "symphony", "drama", "sculpture")
A4 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
B4 <- c("female", "woman", "girl", "sister", "she", "her", "hers", "daughter")
sw <- query(glove_math, S4, T4, A4, B4)

calculate_es(sw)
#> [1] 1.055015
calculate_es(sw, standardize = FALSE)
#> [1] 0.02486533
calculate_es(sw, r = TRUE)
#> [1] 0.4912066
```
