### 2022-01-29 b1a19e7
Add links to papers in all help files fix #22
