+++ R/rnsb.R
@@ -38,7 +38,7 @@
 #' plot_bias(garg_f1)
 #' @author Chung-hong Chan
 #' @references
-#' Sweeney, C., & Najafian, M. (2019, July). A transparent framework for evaluating unintended demographic bias in word embeddings. In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
+#' Sweeney, C., & Najafian, M. (2019, July). [A transparent framework for evaluating unintended demographic bias in word embeddings.](https://aclanthology.org/P19-1162/) In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
 #' @export
 rnsb <- function(w, S_words, A_words, B_words, levels = 1, verbose = FALSE) {
     ## Cleaning
@@ -77,7 +77,7 @@ rnsb <- function(w, S_words, A_words, B_words, levels = 1, verbose = FALSE) {
 #' @return the Kullback-Leibler divergence.
 #' @author Chung-hong Chan
 #' @references
-#' Sweeney, C., & Najafian, M. (2019, July). A transparent framework for evaluating unintended demographic bias in word embeddings. In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
+#' Sweeney, C., & Najafian, M. (2019, July). [A transparent framework for evaluating unintended demographic bias in word embeddings.](https://aclanthology.org/P19-1162/) In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
 #' @export
 rnsb_es <- function(x) {
     if (!"rnsb" %in% class(x)) {
