+++ R/ect.R
@@ -39,7 +39,7 @@
 #' plot_ect(garg_f1)
 #' @author Chung-hong Chan
 #' @references
-#' Dev, S., & Phillips, J. (2019, April). Attenuating bias in word vectors. In The 22nd International Conference on Artificial Intelligence and Statistics (pp. 879-887). PMLR.
+#' Dev, S., & Phillips, J. (2019, April). [Attenuating bias in word vectors.](https://proceedings.mlr.press/v89/dev19a.html) In The 22nd International Conference on Artificial Intelligence and Statistics (pp. 879-887). PMLR. 
 #' @export
 ect <- function(w, S_words, A_words, B_words, verbose = FALSE) {
     w_lab <- rownames(w)
@@ -67,7 +67,7 @@ ect <- function(w, S_words, A_words, B_words, verbose = FALSE) {
 #' @return Spearman Coefficient
 #' @author Chung-hong Chan
 #' @references
-#' Dev, S., & Phillips, J. (2019, April). Attenuating bias in word vectors. In The 22nd International Conference on Artificial Intelligence and Statistics (pp. 879-887). PMLR.
+#' Dev, S., & Phillips, J. (2019, April). [Attenuating bias in word vectors.](https://proceedings.mlr.press/v89/dev19a.html) In The 22nd International Conference on Artificial Intelligence and Statistics (pp. 879-887). PMLR.
 #' @export
 ect_es <- function(x) {
     if (!"ect" %in% class(x)) {
