+++ man/rnsb_es.Rd
@@ -16,7 +16,7 @@ the Kullback-Leibler divergence.
 This function calculates the Kullback-Leibler divergence of the predicted negative probabilities, P, from the uniform distribution.
 }
 \references{
-Sweeney, C., & Najafian, M. (2019, July). A transparent framework for evaluating unintended demographic bias in word embeddings. In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
+Sweeney, C., & Najafian, M. (2019, July). \href{https://aclanthology.org/P19-1162/}{A transparent framework for evaluating unintended demographic bias in word embeddings.} In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
 }
 \author{
 Chung-hong Chan

[…]
