### 2022-01-29 c218882
Make "guess" explicit in the help file fix #21
