### 2022-01-29 a3641e4
Provide an overview of effect sizes in `?calculate_es` fix #23
