### 2022-01-30 a604168
Update README concerning installation instr. fix #11
