+++ README.Rmd
@@ -33,13 +33,13 @@ If your goal is to reproduce the analysis in Caliskan et al (2017), please consi
 
 ## Installation
 
-Recommended: You can install the Github version of sweater with:
+Recommended: install the latest development version
 
 ``` r
-devtools::install_github("chainsawriot/sweater")
+remotes::install_github("chainsawriot/sweater")
 ```
 
-Or the slightly outdated version from CRAN
+or the "stable" release
 
 ```r
 install.packages("sweater")
