### 2022-01-30 49a729c
Update README, add a dedicated "get help" section, ref #18
