### 2022-01-31 f7fa45c
Update documentation, fix #19
to make suggestions about using `query` and `calculate_es` instead of
those individual functions
