+++ R/nas.R
@@ -14,7 +14,7 @@
 
 #' Calculate Normalized Association Score
 #'
-#' This functions quantifies the bias in a set of word embeddings by Caliskan et al (2017). In comparison to WEAT introduced in the same paper, this method is more suitable for continuous ground truth data. See Figure 1 and Figure 2 of the original paper.
+#' This functions quantifies the bias in a set of word embeddings by Caliskan et al (2017). In comparison to WEAT introduced in the same paper, this method is more suitable for continuous ground truth data. See Figure 1 and Figure 2 of the original paper. If possible, please use [query()] instead.
 #' @inheritParams weat
 #' @author Chung-hong Chan
 #' @return A list with class \code{"nas"} containing the following components:
