+++ man/mac.Rd
@@ -7,7 +7,7 @@
 mac(w, S_words, A_words, verbose = FALSE)
 }
 \arguments{
-\item{w}{a numeric matrix of word embeddings (e.g. from rsparse::GloVe)}
+\item{w}{a numeric matrix of word embeddings (e.g. from \code{\link[rsparse:GloVe]{rsparse::GloVe()}})}
 
 \item{S_words}{a character vector of the first set of target words. In an example of studying gender stereotype, it can include occupations such as programmer, engineer, scientists...}
 
@@ -16,7 +16,7 @@ mac(w, S_words, A_words, verbose = FALSE)
 \item{verbose}{logical, whether to display information}
 }
 \value{
-A list with class \code{"rnd"} containing the following components:
+A list with class \code{"mac"} containing the following components:
 \describe{
 \item{\code{$P}}{a vector of cosine similarity values for every word in S_words}
 \item{\code{$S_words}}{the input S_words}
@@ -25,7 +25,7 @@ A list with class \code{"rnd"} containing the following components:
 \code{\link{mac_es}} can be used to obtain the effect size of the test.
 }
 \description{
-This function calculates the mean average cosine similarity (MAC) score proposed in Manzini et al (2019).
+This function calculates the mean average cosine similarity (MAC) score proposed in Manzini et al (2019). If possible, please use \code{\link[=query]{query()}} instead.
 }
 \examples{
 data(googlenews)

[…]
