### 2022-02-02 c4fa5ce
Add clarification of the query section (paper), ref #13
