### 2022-02-02 chainsawriot/sweater#15 @chainsawriot
If you are talking about the paper, I've cited both @mikolov2013distributed (the googlenews word2vec embeddings) and @pennington:2014:G (the GloVE embeddings). The `small_reddit` is not used.

But I can in the paper list out all the sample (but abridged) word embeddings provided in the package.
