### 2022-02-02 0236927
List out and cite all embeddings provided by the package (paper), ref #15
