### 2022-02-02 chainsawriot/sweater#13 @chainsawriot
@cmaimone Thanks for the comments. I've rectified most of your points by explaining the seed words, why some words can be used as target words, etc in the `paper` branch.

c4fa5ceddf42ade4d691640b436f86b545b84861

Regarding the STAB notation, I believe @brunet2019understanding select the four characters out of convenience (similar to the choices by Calikskan et al., XYAB; or V_m, V_1, and V_2 by Garg et al.) and there's no explanation on **why** those four characters were chosen.

![image](https://user-images.githubusercontent.com/637269/152167614-b37016a8-c439-4b8c-b32e-578dd0e32918.png) <br><i>The section  in @brunet2019understanding about the notion</i><p>

I can use some more descriptive names (such as `target_words_1`, `target_words_2`, `attribute_words_1`, `attribute_words_2`). But again, it breaks legacy code. And I think the documentation, the README, and now the paper have done a decent job in explaining what those 4 word sets are.
