### 2022-02-02 6d4eb04
Improve the flow of the statement of need section (paper), ref #17
