### 2022-02-02 edc7cdb
Improve the paper and the benchmark ref #16, #13, #15, #17 (#33)
* Add related work section, fix #16

* suggested clarifications (#12)

on acceptable word embedding formats

re: https://github.com/openjournals/joss-reviews/issues/4036

* backport the clarification to the `rmd`

* Add python comparision in benchmark

* Add clarification of the query section (paper), ref #13

* List out and cite all embeddings provided by the package (paper), ref #15

* Improve the flow of the statement of need section (paper), ref #17

Co-authored-by: Christina Maimone <christina.maimone@gmail.com>
