+++ paper/paper.bib
@@ -234,3 +234,13 @@
   year={2019},
   organization={PMLR}
 }
+
+@article{collobert2011natural,
+  title={Natural language processing (almost) from scratch},
+  author={Collobert, Ronan and Weston, Jason and Bottou, L{\'e}on and Karlen, Michael and Kavukcuoglu, Koray and Kuksa, Pavel},
+  journal={Journal of machine learning research},
+  volume={12},
+  number={ARTICLE},
+  pages={2493--2537},
+  year={2011}
+}
