+++ paper/bench.R
@@ -1,0 +1,12 @@
+require(sweater)
+
+big_glove <- read_word2vec("../raw_data/glove.840B.300d.txt")
+
+S2 <- c("math", "algebra", "geometry", "calculus", "equations",
+        "computation", "numbers", "addition")
+T2 <- c("poetry", "art", "dance", "literature", "novel", "symphony",
+        "drama", "sculpture")
+A2 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
+B2 <- c("female", "woman", "girl", "sister", "she", "her", "hers",
+        "daughter")
+query(big_glove, S2, T2, A2, B2, method = "weat")

[…]
