### 2022-02-02 chainsawriot/sweater#13 @cmaimone
I wasn't suggesting renaming them - just maybe making it clearer what each is, since they are abstract letters.  I figured it out eventually, but I got lost on my first read through the paper/example code.

For someone who hasn't read the original papers (I have intentionally not done so at this point since I think others will come to your package without having done so), the query section was hard to work through.  It was clearer once I had an example.  Maybe a brief example before the usage section would help -- not necessarily in code.  It would help to know what we're doing before knowing the functions/code to do it.  Then you could map the STAB letters back to the example.

This would, I think, also help people who have read some of these original papers, yes?  Because the STAB notation is not consistent across them?
