### 2022-02-02 0c7716e
Correct exact test; add vignette
