+++ R/RcppExports.R
@@ -21,7 +21,3 @@ cpp_nas <- function(c, A, glove_mat) {
     .Call(`_sweater_cpp_nas`, c, A, glove_mat)
 }
 
-cpp_exact <- function(union_diff, test_stat, s_length) {
-    .Call(`_sweater_cpp_exact`, union_diff, test_stat, s_length)
-}
-
