+++ vignettes/resampling.Rmd
@@ -1,0 +1,123 @@
+---
+title: "resampling"
+output: rmarkdown::html_vignette
+
+vignette: >
+  %\VignetteIndexEntry{resampling}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The original significance test proposed by Caliskan et al. in their paper is based on exact inference using a permutaion approach. This package provides such an exact test. But like any permutation test (such as Fisher's exact test), the number of permuations, as well as the running time, grows exponentially with the number of samples (in the case, the total length of S and T).
+
+In the following benchmark, we studied how much time does it take to finish an exact test with the word set with a length of 2 to 4. I actually don't recommend trying anything beyond 6; it would take probably a few hours, days even. And actually, the function `weat_exact` will warn you about that.
+
+```{r setup}
+library(sweater)
+S1 <- c("math", "algebra", "geometry", "calculus", "equations",
+ "computation", "numbers", "addition")
+T1 <- c("poetry", "art", "dance", "literature", "novel", "symphony", "drama", "sculpture")
+A1 <- c("male", "man", "boy", "brother", "he", "him", "his", "son")
+B1 <- c("female", "woman", "girl", "sister", "she", "her", "hers", "daughter")
+
+do_exact_weat <- function(n = 4) {
+    sw <- weat(glove_math, S1[1:n], T1[1:n], A1[1:n], B1[1:n])
+    weat_exact(sw)
+}
+
+require(bench)
+benchmark_res <- bench::mark(do_exact_weat(2),
+                             do_exact_weat(3),
+                             do_exact_weat(4),
+                             relative = TRUE)
+benchmark_res
+```
+
[…]
