### 2022-02-02 chainsawriot/sweater#14 @chainsawriot
A Vignette has been added to explain the resampling test.

https://github.com/chainsawriot/sweater/blob/master/vignettes/resampling.Rmd
