### 2022-02-03 #4036 @chainsawriot
Dear @cmaimone @kbenoit @sbenthall

I would like to express my sincere gratitude to @cmaimone for her comments about both the paper and the package. Her comments have leaded to significant improvement of the package.

* chainsawriot/sweater#27 : `S_diff` and `T_diff` from the S3 object `weat` are now named vectors.
* chainsawriot/sweater#26 : The error message of the plot function has been improved.
* chainsawriot/sweater#23 : Ranges of all effect sizes and their interpretations have been added to the documentation of the function `calculate_es`.
* chainsawriot/sweater#22 : All cited papers in the documentation are now clickable.
* chainsawriot/sweater#21 : The "guess" parameter of `query` is explained in the README as well as in the documentation.
* chainsawriot/sweater#20 : The first example in the README has been now elaborated in detail.
* chainsawriot/sweater#19 : The preferred workflow is now elaborated in the README as well as in the documentation
* chainsawriot/sweater#18 : A get help section has been added to the README.
* chainsawriot/sweater#17 : The 1st paragraph of the paper has been rewritten to provide more context as well as more examples before elaborating on the social science theory.
* chainsawriot/sweater#16 : A related work section has been added to the paper.
* chainsawriot/sweater#15 : All datasets provided by the package have been listed out and cited in the paper.
* chainsawriot/sweater#14 : See below. But a vignette has been added to show the accuracy of the resampling approximation to the permutation exact inference.
* chainsawriot/sweater#13 : The Query section in the paper has been rewritten to elaborate better what those STAB word sets are.
* chainsawriot/sweater#11 : The installation instruction has been updated to recommend using `remotes` as well as clarify the CRAN version is "stable" release.

I think I've addressed most of her comments. There is still one hanging issue and that's the speed claim. I've added a benchmark of the package versus a native implementation in R as well as versus the Python `wefe`. chainsawriot/sweater#14 The workflow provided by `sweater` is still at least 10 times faster.

I've an idea of adding the Java code (provided by Caliskan et al.). But I have only limited knowledge in Java and I need more time. In the mean time, I think I would benefit from the comment by @kbenoit regarding the speed issue. I would like to know how @kbenoit thinks about the speed issue.

Regards,
Chung-hong Chan
