### 2022-02-03 0c4b585
Correct the time complexity of the permutation test
