### 2022-02-09 a7903e2
Add Java benchmark ref #14
