### 2022-02-09 c8d81f5
Update java steps
