+++ paper/benchmark.md
@@ -133,9 +133,9 @@ benchmark_res
     ## # A tibble: 3 × 6
     ##   expression                             min median `itr/sec` mem_alloc `gc/sec`
     ##   <bch:expr>                           <dbl>  <dbl>     <dbl>     <dbl>    <dbl>
-    ## 1 r_weat(glove_math, S2, T2, A2, B2)    14.3   13.1      1         2.35     1   
-    ## 2 r_weat_c(glove_math, S2, T2, A2, B2)  13.7   13.1      1.06      2.35     1.15
-    ## 3 cpp_weat(glove_math, S2, T2, A2, B2)   1      1       13.5       1        2.50
+    ## 1 r_weat(glove_math, S2, T2, A2, B2)    13.5   12.2      1.12      2.35     1   
+    ## 2 r_weat_c(glove_math, S2, T2, A2, B2)  13.5   12.6      1         2.35     1.02
+    ## 3 cpp_weat(glove_math, S2, T2, A2, B2)   1      1       13.9       1        2.54
 
 ### Random benchmark
 
@@ -181,16 +181,16 @@ res %>% map_dfr(~.[1,3]) %>% dplyr::mutate(stab_length = stab_length)
     ## # A tibble: 10 × 2
     ##    median stab_length
     ##     <dbl>       <dbl>
-    ##  1  10.8           10
-    ##  2  10.1           20
-    ##  3   9.24          30
-    ##  4  11.4           40
-    ##  5   9.37          50
-    ##  6   9.28          60
-    ##  7   9.37          70
-    ##  8   9.30          80
-    ##  9   8.84          90
-    ## 10   9.07         100
+    ##  1   9.93          10
+    ##  2  11.4           20
+    ##  3  11.6           30
+    ##  4   7.67          40
+    ##  5  12.0           50
+    ##  6   9.10          60
+    ##  7   9.60          70
+    ##  8   8.54          80
+    ##  9   9.94          90
+    ## 10   9.16         100
 
 ### versus WEFE
 
@@ -219,9 +219,9 @@ time Rscript bench.R
     ## • `calculate_es()`: Calculate effect size
     ## • `weat_resampling()`: Conduct statistical test
     ## 
-    ## real 0m25.643s
-    ## user 1m0.617s
-    ## sys  0m7.542s
+    ## real 0m23.101s
+    ## user 0m55.114s
+    ## sys  0m6.360s
 
 The Python workflow, however, needs to use `gensim` to read the
 pretained word embedding file and it can’t read GLoVE format directly
@@ -237,15 +237,16 @@ time python3 bench.py
 
     ## {'query_name': 'S and T wrt A and B', 'result': 0.19892264262307435, 'weat': 0.19892264262307435, 'effect_size': 1.0896144272748516, 'p_value': nan}
     ## 
-    ## real 14m7.488s
-    ## user 13m47.748s
-    ## sys  0m20.704s
+    ## real 12m6.472s
+    ## user 11m46.732s
+    ## sys  0m19.764s
 
 ### versus the original Java code
 
 The original Java code by Caliskan et al. is extremely fast because the
 code is highly optimized. If all you need to do is WEAT and you know how
-to write Java, it is recommended using the Java code.
+to write Java, it is recommended using the Java code. See this footnote
+\[2\] on how to run the code.
 
 ``` bash
 ## javac -cp ./lib/commons-lang3-3.3.2.jar:./lib/commons-math3-3.6.1.jar WeatBenchmark.java Utils.java
[…]
