### 2022-03-25 #4036 @kbenoit
Sorry for the delay, I won't make excuses, but they involved staffing issues at the Institute I direct, and me getting COVID.

This is a very useful and impressively implemented package. I think the rationale is clearly explained, and the package documentation makes it very clear how and when it can be used and for what purposes. It installs easily and runs very quickly.

I am in favour of acceptance, but it would be nice to see some of the issues addressed below.

Two overall comments, in addition to the specific ones below. The title of the package relates to association tests, but the discussion almost instantly focuses on bias. Yet I can also see the association being useful for less pejorative concepts than bias, for instance political affiliations, such as the ability to test target words' associations with say left- or right-leaning political parties. Why not state this in the conclusion, that there could be uses of the association tests beyond seeking bias?

Second, I love (of course) the support for **quanteda** dictionaries, since that's an aspect of the package that @koheiw and I have spent a very large amount of time developing. The README has some details but it would be great - and relatively easy - to extend this as described as "coming soon" in the README, and make a note of this in the article.

The Bing dictionaries are implemented in `data_dictionary_HuLiu` in https://github.com/quanteda/quanteda.sentiment, which is still not on CRAN but working. **quanteda** dictionaries are just fancy named lists, so it would be easy to write methods to make them work as A_words or B_words too, and to extend this to `rnd()` and `weat()`. It would be nice to have a mention of this in the _JOSS_ article, if possible, but not a deal-breaker for me.

I look forward to using the package!

From the article:
[ ] "only scatteredly available" - perhaps reword to "only available piecemeal". In that paragraph I'd also suggest revising the last sentence to make it clear that it's **sweater** that provides the Rcpp implementations. (The current passive voice makes this unclear.)
[ ] Because `LazyData` is `true`, there is no need to use e.g. `data(glove_math)` in the examples.
