### 2022-03-25 #4036 @cmaimone
@chainsawriot - I'm trying to finish off my checklist now that the other review is complete -- is the paper in the master branch up to date?  If so, can you regenerate the pdf here?

I also second @kbenoit points about association vs. bias -- I think addressing this would help with some of the issues in the paper I noted as well.  Note that it's also in the documentation/examples:

> All tests in this package use the concept of queries (see Badilla et al., 2020) to study the biases in the input word embeddings w.

It's not just bias, right?  It's association
