### 2022-03-28 #4036 @chainsawriot
I would like to thank @kbenoit for his comments. I am sorry to learn that you caught COVID. I wish you good health and be as energetic as a dragon and a tiger (a Cantonese proverb).

I will deal with other points later. But the point about bias has also been raised by @cmaimone now, maybe I should ask for your opinion on how to deal with this.

I agree with both of you that the so-called (social) bias detection methods are actually finding differential (implicit) associations between target words and attribute words. It is similar to the Harvard [Implicit Association Test]([https://doi.org/10.1037/0022-3514.74.6.1464](https://doi.apa.org/doi/10.1037/0022-3514.74.6.1464)): Although IAT is exclusively used to test for human biases, it is not called Implicit Bias Test. I also agree with @kbenoit that one can use the methods provided by sweater to measure things in word embeddings that are not biases. I don't aware of a definition of word embedding bias. To me, if a target word should be rightfully associated with an attribute and such association is a wanted one (e.g. "Diamond" -> "forever" vs "temporary"; "Angela Merkel" -> "German" vs "British") , then such association is not a bias.

In the computer science / social science literature, researchers are interested in detecting unwanted associations (e.g. racial / gender stereotypes) and thus are biases.

I was wondering should I put it this way:

*The goal of this R package is to detect associations among words in word embeddings spaces. Word embeddings can capture how similar or different two words are in terms of implicit and explicit meanings. Using the example in @collobert2011natural, the word vector for "XBox" is close to that of "PlayStation", as measured by a distance measure such as cosine distance. The same technique can also be used to detect unwanted implicit associations, or biases. In the situation of racial bias detection, for example, @kroon2020guilty measure how close the word vectors for various ethnic group names (e.g. "Dutch", "Belgian" , and "Syrian") to that of various nouns related to threats (e.g. "terrorist", "murderer", and "gangster"). These biases in word embedding can be understood through the implicit social cognition model of media priming [@arendt:2013:DDM]. In this model, implicit stereotypes are defined as the "strength of the automatic association between a group concept (e.g., minority group) and an attribute (e.g., criminal)." [@arendt:2013:DDM, p. 832] All of these detection methods are based on the strength of association between a concept (or a target) and an attribute in embedding spaces.*
