### 2022-03-28 #4036 @cmaimone
I think I closed out the issues I opened.  One PR open for small wording clarifications on the most recent copy of the paper.  I'm good with accepting and publishing now
