### 2022-03-29 cb391a5
small wording edits for clarity (#35)
* small wording edits for clarity

* Knit
