### 2022-03-29 5892b12
Update the JOSS paper according to @kbenoit's comments
