+++ paper/paper.bib
@@ -244,3 +244,26 @@
   pages={2493--2537},
   year={2011}
 }
+
+@article{jing2021characterizing,
+  title={Characterizing partisan political narrative frameworks about {COVID-19} on {Twitter}},
+  author={Jing, Elise and Ahn, Yong-Yeol},
+  journal={EPJ data science},
+  volume={10},
+  number={1},
+  pages={53},
+  year={2021},
+  publisher={Springer Berlin Heidelberg},
+  doi={10.1140/epjds/s13688-021-00308-4}
+}
+
+@article{benoit2018quanteda,
+  title={quanteda: An {R} package for the quantitative analysis of textual data},
+  author={Benoit, Kenneth and Watanabe, Kohei and Wang, Haiyan and Nulty, Paul and Obeng, Adam and M{\"u}ller, Stefan and Matsuo, Akitaka},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={30},
+  pages={774},
+  year={2018},
+  doi={10.21105/joss.00774}
+}
