### 2022-03-30 a81bf8a
Update README to reflect the changes in the JOSS paper
