### 2022-03-30 #4036 @chainsawriot
@kbenoit

Regarding your comments, I have made the following changes

* Biases / Associations - I have made it clear that the package provides methods for detecting associations; and wanted associations are biases. A published article (Jing & Ahn, 2021) has been cited as an example of the application you suggested. The README is also revised to have a consistent framing.

https://github.com/chainsawriot/sweater/blob/a81bf8ab812535c38e1164afd1aff0aeb0179b53/paper/paper.rmd#L31

* Support for quanteda dictionaries - Currently only the method `rnsb` supports quanteda dictionaries. And yes, it is not that difficult to make other methods support dictionaries. Extending the support is on my schedule. But at the moment, `rnsb`'s dictionary support is well tested. I have taken your advice to hint the future support of quanteda dictionaries in both the paper and the README.

https://github.com/chainsawriot/sweater/blob/a81bf8ab812535c38e1164afd1aff0aeb0179b53/paper/paper.rmd#L87

The minor points

* I have taken your advice and revised the sentences as you suggested.

https://github.com/chainsawriot/sweater/blob/a81bf8ab812535c38e1164afd1aff0aeb0179b53/paper/paper.rmd#L37

* I have removed all `data()` calls in all examples

I would like to thank @cmaimone for the pull request as well!

@sbenthall Please let me know what I should do next. Thank you very much!
