### 2022-03-31 #4036 @sbenthall
@chainsawriot A few more comments on the paper based on my read-through:

> All methods require S while T is only required for WEAT.

This sentence occurs before you say what WEAT is. It would be better to give the meaning of the acronym the first time you mention it.
