### 2022-03-31 ca5c9d0
Update paper: missing DOI and expand WEAT openjournals/joss-reviews/#4036
