### 2022-04-07 #4036 @sbenthall
I recommend this submission for acceptance to JOSS.
