### 2022-04-10 #4036 @arfon
@kbenoit, @cmaimone  – many thanks for your reviews here and to @sbenthall for editing this submission. This review looks like a model JOSS experience – thanks to all for making a collaborative, constructive review experience! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
